// src/app/interfaces/car.ts
export interface Car {
    id: number;
    codigo: string;
    marca: string;
    modelo: string;
    foto: string;
    anio: number;
    calificacion: number;
    fecha_creacion: any;
    fecha_modificacion: any;
    usuario: string;
    usuario_mod: string;
    kilometraje: number;
    precio: string;
}